test: split Z2M tests into test_e2e, test_provider, test_payload modules
Entire-Checkpoint: 9df4875153b9
