+++ tests/providers/zigbee2mqtt/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the Zigbee2MQTT lock provider."""
